feat: 2단계 요구사항에 맞게 구조 변경 (지하철 구간 제외)
