+++ src/main/java/nextstep/subway/domain/Line.java
@@ -17,17 +17,28 @@ public class Line extends BaseEntity {
     @Column(name = "color")
     private String color;
 
-    @OneToMany(mappedBy = "line")
-    private List<Section> Sections = new ArrayList<>();
+    @OneToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "UP_STATION_ID")
+    private Station upStation;
+
+    @OneToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "DOWN_STATION_ID")
+    private Station downStation;
 
     protected Line() {
 
     }
 
     public Line(String name, String color) {
+        this(name, color, null, null);
+    }
+
+    public Line(String name, String color, Station upStation, Station downStation) {
         this.id = null;
         this.name = name;
         this.color = color;
+        this.upStation = upStation;
+        this.downStation = downStation;
     }
 
     public Long getId() {
@@ -42,8 +53,12 @@ public class Line extends BaseEntity {
         return color;
     }
 
-    public List<Section> getSections() {
-        return Sections;
+    public Station getUpStation() {
+        return upStation;
+    }
+
+    public Station getDownStation() {
+        return downStation;
     }
 
     @Override
@@ -52,12 +67,12 @@ public class Line extends BaseEntity {
         if (!(o instanceof Line)) return false;
         Line line = (Line) o;
         if (getId() != null && Objects.equals(getId(), line.getId())) return true;
-        return Objects.equals(getId(), line.getId()) && Objects.equals(getName(), line.getName()) && Objects.equals(getColor(), line.getColor()) && Objects.equals(getSections(), line.getSections());
+        return Objects.equals(getId(), line.getId()) && Objects.equals(getName(), line.getName()) && Objects.equals(getColor(), line.getColor()) && Objects.equals(getUpStation(), line.getUpStation()) && Objects.equals(getDownStation(), line.getDownStation());
     }
 
     @Override
     public int hashCode() {
-        return Objects.hash(getId(), getName(), getColor(), getSections());
+        return Objects.hash(getId(), getName(), getColor(), getUpStation(), getDownStation());
     }
 
     @Override
